# firmware: optimize integer array/list rpc