# ARTIQ Pull Request

## Description of Changes
Optimize integer array/list device-to-host RPC.

As the device is in big endian, we can cast the slice into byte slice, directly copy from slice instead of calling `write_u32`/`write_u64` repeatedly. This improves the performance quite a lot.

| Test                 | Prev (MiB/s) | Now (MiB/s)  |  diff  |
| -------------------- | ------------ | ------------ | ------ |
| I32 Array (1MB) H2D  |         2.02 |         2.06 | +1.90% |
| I32 Array (1MB) D2H  |         0.48 |         2.23 |  +365% |
| I32 Array (1KB) H2D  |         0.36 |         0.35 | -2.78% |
| I32 Array (1KB) D2H  |         0.23 |         0.38 | +65.2% |
| I32 List (1MB) H2D   |         2.07 |         2.00 | -3.38% |
| I32 List (1MB) D2H   |         0.45 |         2.18 |  +384% |
| I32 List (1KB) H2D   |         0.35 |         0.35 | -0.00% |
| I32 List (1KB) D2H   |         0.23 |         0.39 | +69.6% |

### Related Issue

Closes #1255

## Type of Changes

<!-- Leave ONLY the corresponding lines for the applicable type of change: -->
|   | Type |
| ------------- | ------------- |
| ✓  | :hammer: Refactoring  |

## Steps (Choose relevant, delete irrelevant before submitting)

### All Pull Requests

- [x] Use correct spelling and grammar.
- [ ] Close/update issues.

### Code Changes

- [x] Test your changes or have someone test them. Mention what was tested and how.
- [x] Add and check docstrings and comments

### Git Logistics

- [x] Split your contribution into logically separate changes (`git rebase --interactive`). Merge/squash/fixup commits that just fix or amend previous commits. Remove unintended changes & cleanup. See [tutorial](https://www.atlassian.com/git/tutorials/rewriting-history/git-rebase).
- [x] Write short & meaningful commit messages. Review each commit for messages (`git show`). Format:
  ```
  topic: description. < 50 characters total.
  
  Longer description. < 70 characters per line
  ```

### Licensing

See [copyright & licensing for more info](https://github.com/m-labs/artiq/blob/master/CONTRIBUTING.rst#copyright-and-sign-off).
ARTIQ files that do not contain a license header are copyrighted by M-Labs Limited and are licensed under LGPLv3+.