@dnadlinger 2021-01-21
The improvements in device-to-host transfers for integers are really useful, as that's what most of the data tends to be (e.g. input event timestamps).